Write segfault

+++ src/converters.cpp
@@ -52,6 +52,7 @@ std::wstring CONVERT::wstr(const std::string &src)
 std::string CONVERT::convert_from_bytes_to_string(const std::vector<uint8_t> &from)
 {
     std::string str;
+   // std::cout<<"\n55";
     size_t size = from.size();
     str.reserve(size);
     for (size_t i = 0; i < size; ++i)
@@ -81,7 +82,7 @@ std::string &CONVERT::low(std::string &str)
 std::string CONVERT::to_hex(uint8_t *response, size_t size)
 {
     std::ostringstream oss;
-    for (int i = 0; i < size; i++)
+    for (size_t i = 0; i < size; i++)
         oss << std::hex << std::setw(2) << std::setfill('0') << static_cast<int>(response[i]) << " ";
     return oss.str();
 }
